requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
matplotlib>=3.7
seaborn>=0.12
fastapi>=0.104
uvicorn>=0.24
pydantic>=2.0
sqlalchemy>=2.0
pyyaml>=6.0
pytest>=7.4
pytest-cov>=4.1
black>=23.0
isort>=5.12
flake8>=6.1
joblib>=1.3
httpx>=0.25
python-dotenv>=1.0
shap>=0.43
mlflow>=2.8
evidently>=0.4
optuna>=3.4
imbalanced-learn>=0.11
ucimlrepo>=0.0.3
